requirements.txt
Django==4.2.19
requests
whitenoise
dj-database-url
psycopg[binary]
python-decouple
gunicorn
django-htmx
psycopg2
django-compressor
cloudinary
pillow

